-new Article table seeder,
-adding call for seed the article table
-modify article table field definition

+++ database/migrations/2017_02_01_161414_create_articles_table.php
@@ -19,7 +19,7 @@ class CreateArticlesTable extends Migration
             $table->string('descripcion');
             $table->integer('stock')->nullable(false);
             $table->integer('category_id')->nullable(false);
-            $table->decimal('precio_unitario',10,2);
+            $table->float('precio_unitario',8,2);
             $table->timestamps();
         });
     }
